requirement.txt
streamlit==1.1.0
qrcode==7.3
Pillow==8.3.2
